Se suben validaciones que no validan :P

+++ TallerMecanico/src/servicios/AutoBusinessObjectImpl.java
@@ -34,24 +34,39 @@ public class AutoBusinessObjectImpl implements AutoBusinessObject {
 	@Override
 	public void insertAuto(Auto auto)
 			throws TallerMecanicoException, ExistingCarException, NoIdObtainedException, CamposVaciosException {
-		if (auto.getPatente() != null && auto.getMarca() != null && auto.getModelo() != null && auto.getColor() != null
-				&& auto.getCantidadPuertas() != 0 && auto.getAño() != null && auto.getKilometraje() != 0) {
-			autoDAO.insertAuto(auto);
+		if (auto.getPatente().equals("") && auto.getMarca().equals("") && auto.getModelo().equals("")
+				&& auto.getColor().equals("") && auto.getCantidadPuertas() != 0 && auto.getAño().equals("")
+				&& auto.getKilometraje() != 0) {
+			if (autoDAO.existeAuto(auto.getPatente())) {
+				throw new ExistingCarException("El auto que intenta dar de alta ya existe");
+			} else {
+				autoDAO.insertAuto(auto);
+			}
 		} else {
 			throw new CamposVaciosException("No puede haber espacios sin completar");
 		}
 	}
 
 	@Override
-	public void updateAuto(Auto auto) throws TallerMecanicoException, NonExistingCarException {
-		autoDAO.updateAuto(auto);
+	public void updateAuto(Auto auto) throws TallerMecanicoException, NonExistingCarException, CamposVaciosException {
+		if (auto.getPatente().equals("") && auto.getMarca().equals("") && auto.getModelo().equals("")
+				&& auto.getColor().equals("") && auto.getCantidadPuertas() != 0 && auto.getAño().equals("")
+				&& auto.getKilometraje() != 0) {
+			autoDAO.updateAuto(auto);
+		} else {
+			throw new CamposVaciosException("No puede haber espacios sin completar");
+		}
 	}
 
 	@Override
 	public void deleteAuto(String patente)
 			throws TallerMecanicoException, NonExistingCarException, CamposVaciosException {
 		if (!patente.equals("")) {
-			autoDAO.deleteAuto(patente);
+			if (autoDAO.existeAuto(patente)) {
+				autoDAO.deleteAuto(patente);
+			} else {
+				throw new NonExistingCarException("El auto que intenta modificar no existe");
+			}
 		} else {
 			throw new CamposVaciosException("No puede haber espacios sin completar");
 		}
